requirements.txt
langgraph 
langchain 
openai
zep-python 
pinecone-client 
weaviate-client 
neo4j
python-dotenv
ollama
pyodbc
reportlab
datetime
numpy